тесты

+++ src/main/java/by/toukachmikhail/taskmanagementsystem/services/impl/TaskServiceImpl.java
@@ -2,9 +2,9 @@ package by.toukachmikhail.taskmanagementsystem.services.impl;
 
 import static by.toukachmikhail.taskmanagementsystem.exception_handling.enums.AccessDeniedExceptionMessage.ACCESS_DENIED_EXCEPTION_MESSAGE;
 import static by.toukachmikhail.taskmanagementsystem.exception_handling.enums.NotFoundExceptionMessage.ASSIGNEE_NOT_FOUND;
-import static by.toukachmikhail.taskmanagementsystem.exception_handling.enums.NotFoundExceptionMessage.USER_NOT_FOUND;
 import static by.toukachmikhail.taskmanagementsystem.exception_handling.enums.NotFoundExceptionMessage.TASK_NOT_FOUND;
 
+import by.toukachmikhail.taskmanagementsystem.dto.CommentDto;
 import by.toukachmikhail.taskmanagementsystem.dto.TaskDto;
 import by.toukachmikhail.taskmanagementsystem.entities.Comment;
 import by.toukachmikhail.taskmanagementsystem.entities.Task;
@@ -14,6 +14,7 @@ import by.toukachmikhail.taskmanagementsystem.exception_handling.exception.Acces
 import by.toukachmikhail.taskmanagementsystem.exception_handling.exception.NotFoundException;
 import by.toukachmikhail.taskmanagementsystem.mappers.CommentMapper;
 import by.toukachmikhail.taskmanagementsystem.mappers.TaskMapper;
+import by.toukachmikhail.taskmanagementsystem.mappers.UserMapper;
 import by.toukachmikhail.taskmanagementsystem.repositories.CommentRepository;
 import by.toukachmikhail.taskmanagementsystem.repositories.TaskRepository;
 import by.toukachmikhail.taskmanagementsystem.repositories.UserRepository;
@@ -37,39 +38,31 @@ public class TaskServiceImpl implements TaskService {
   private final CommentRepository commentRepository;
   private final TaskMapper taskMapper;
   private final CommentMapper commentMapper;
+  private final UserDetailsServiceImpl userDetailsService;
+  private final UserMapper userMapper;
 
   @Override
   public Page<TaskDto> getAllTasks(int page, int size, String sortBy, String direction) {
+    User currentUser = userDetailsService.getCurrentUser();
     Sort sort = Sort.by(Sort.Direction.fromString(direction), sortBy);
     Pageable pageable = PageRequest.of(page, size, sort);
-    Page<Task> allTasks = taskRepository.findAll(pageable);
 
-    return allTasks.map(taskMapper::entityToDto);
+    if (currentUser.getRole() == UserRole.USER) {
+      Page<Task> currentUserTasks = taskRepository.findByAssignee(
+          currentUser, pageable);
+      return currentUserTasks.map(taskMapper::entityToDtoForList);
+    } else if (currentUser.getRole() == UserRole.ADMIN) {
+      Page<Task> currentAdminTasks = taskRepository.findByAuthor(currentUser, pageable);
+      return currentAdminTasks.map(taskMapper::entityToDtoForList);
+    }
+    return Page.empty();
   }
 
   @Override
   public TaskDto getTaskById(Long taskId) {
     Task task = taskRepository.findById(taskId)
         .orElseThrow(() -> new NotFoundException(TASK_NOT_FOUND.getMessage()));
-    return taskMapper.entityToDto(task);
-  }
-
-  @Override
-  public Page<TaskDto> getTasksByAuthor(Long authorId, int page, int size, String sortBy,
-      String direction) {
-    Sort sort = Sort.by(Sort.Direction.fromString(direction), sortBy);
-    Pageable pageable = PageRequest.of(page, size, sort);
-    Page<Task> tasks = taskRepository.findByAuthorId(authorId, pageable);
-    return tasks.map(taskMapper::entityToDto);
-  }
-
-  @Override
-  public Page<TaskDto> getTasksByAssignee(Long assigneeId, int page, int size, String sortBy,
-      String direction) {
-    Sort sort = Sort.by(Sort.Direction.fromString(direction), sortBy);
-    Pageable pageable = PageRequest.of(page, size, sort);
-    Page<Task> tasks = taskRepository.findByAuthorId(assigneeId, pageable);
-    return tasks.map(taskMapper::entityToDto);
+    return taskMapper.entityToDtoForList(task);
   }
 
   @Override
@@ -81,15 +74,10 @@ public class TaskServiceImpl implements TaskService {
     task.setStatus(taskDTO.status());
     task.setTaskPriority(taskDTO.priority());
 
-//    User author = userRepository.findByEmail(taskDTO.author().email())
-//        .orElseThrow(() -> new NotFoundException(USER_NOT_FOUND.getMessage()));
     User assignee = userRepository.findByEmail(taskDTO.assignee().email())
         .orElseThrow(() -> new NotFoundException(ASSIGNEE_NOT_FOUND.getMessage()));
 
-    author.setUsername(taskDTO.author().username());
-    author.setEmail(taskDTO.author().email());
-    author.setRole(taskDTO.author().role());
-    assignee.setRole(UserRole.USER);
+    User author = userDetailsService.getCurrentUser();
 
     task.setAuthor(author);
     task.setAssignee(assignee);
@@ -103,63 +91,80 @@ public class TaskServiceImpl implements TaskService {
       comment.setUser(assignee);
 
       commentRepository.save(comment);
+      createdTask.getComments().add(comment);
     }
 
-    return taskMapper.entityToDto(createdTask);
+    return taskMapper.entityToDtoForCreation(createdTask);
   }
 
   @Override
-  public TaskDto updateTask(Long id, TaskDto taskDTO, User currentUser) {
+  public TaskDto updateTask(Long id, TaskDto taskDto, User currentUser) {
     log.info("Method updateTask called for task ID: {}", id);
     log.info("Current user: {}", currentUser.getUsername());
     log.info("Current user role: {}", currentUser.getRole());
 
     Task task = taskRepository.findById(id)
         .orElseThrow(() -> new NotFoundException(TASK_NOT_FOUND.getMessage()));
-    log.info("User {}", currentUser.getRole());
 
     if (currentUser.getRole() == UserRole.USER) {
       if (!task.getAssignee().equals(currentUser)) {
         throw new AccessDeniedException(ACCESS_DENIED_EXCEPTION_MESSAGE.getMessage());
       }
 
-      if (taskDTO.status() != null) {
-        task.setStatus(taskDTO.status());
+      if (taskDto.status() != null) {
+        task.setStatus(taskDto.status());
       }
     } else {
-      if (taskDTO.header() != null) {
-        task.setHeader(taskDTO.header());
+      if (taskDto.header() != null) {
+        task.setHeader(taskDto.header());
       }
-      if (taskDTO.description() != null) {
-        task.setDescription(taskDTO.description());
+      if (taskDto.description() != null) {
+        task.setDescription(taskDto.description());
       }
-      if (taskDTO.status() != null) {
-        task.setStatus(taskDTO.status());
+      if (taskDto.status() != null) {
+        task.setStatus(taskDto.status());
       }
-      if (taskDTO.priority() != null) {
-        task.setTaskPriority(taskDTO.priority());
+      if (taskDto.priority() != null) {
+        task.setTaskPriority(taskDto.priority());
       }
-      if (taskDTO.assignee() != null) {
-        User assignee = userRepository.findByUsername(taskDTO.assignee().username())
+      if (taskDto.assignee() != null) {
+        User assignee = userRepository.findByUsername(taskDto.assignee().username())
             .orElseThrow(() -> new NotFoundException(ASSIGNEE_NOT_FOUND.getMessage()));
         task.setAssignee(assignee);
       }
     }
 
-    if (taskDTO.comment() != null) {
-      Comment newComment = commentMapper.dtoToEntity(taskDTO.comment());
-      newComment.setTask(task);
-      newComment.setUser(currentUser);
-      task.getComments().add(newComment);
+    CommentDto newCommentDto = null;
+    if (taskDto.comment() != null) {
+      Comment comment = commentMapper.dtoToEntity(taskDto.comment());
+      comment.setTask(task);
+      comment.setUser(currentUser);
+      task.getComments().add(comment);
+      commentRepository.save(comment);
+      newCommentDto = commentMapper.entityToDTO(comment);
     }
 
     Task updatedTask = taskRepository.save(task);
-    return taskMapper.entityToDto(updatedTask);
+    return TaskDto.builder()
+        .header(updatedTask.getHeader())
+        .description(updatedTask.getDescription())
+        .status(updatedTask.getStatus())
+        .priority(updatedTask.getTaskPriority())
+        .assignee(userMapper.entityToDto(updatedTask.getAssignee()))
+        .comment(newCommentDto)
+        .build();
   }
 
   @Override
   public void deleteTask(Long taskId) {
+    if (!isTaskExistById(taskId)) {
+      throw new NotFoundException(TASK_NOT_FOUND.getMessage());
+    }
     taskRepository.deleteById(taskId);
+  }
+
+  private boolean isTaskExistById(Long taskId) {
+    return taskRepository.existsById(taskId);
   }
 
   @Override
@@ -168,5 +173,4 @@ public class TaskServiceImpl implements TaskService {
         .orElseThrow(() -> new NotFoundException("Task not found"));
     return task.getAssignee().getUsername().equals(username);
   }
-
 }
